--- a/pkg/proxy/ipvs/proxier.go
+++ b/pkg/proxy/ipvs/proxier.go
@@ -1843,7 +1843,7 @@ func (proxier *Proxier) syncEndpoint(svcPortName proxy.ServicePortName, onlyNode
 	}
 
 	// Create new endpoints
-	for _, ep := range newEndpoints.UnsortedList() {
+	for ep := range newEndpoints {
 		ip, port, err := net.SplitHostPort(ep)
 		if err != nil {
 			proxier.logger.Error(err, "Failed to parse endpoint", "endpoint", ep)