--- a/staging/src/k8s.io/apimachinery/pkg/util/sets/set_test.go
+++ b/staging/src/k8s.io/apimachinery/pkg/util/sets/set_test.go
@@ -284,6 +284,22 @@ func TestStringIntersection(t *testing.T) {
 	}
 }
 
+func TestInsertSet(t *testing.T) {
+	s1 := New("1")
+	s2 := New("2")
+	s1.InsertSet(s2)
+
+	if s1.Len() != 2 {
+		t.Errorf("Expected Len()=2 but got %d", s1.Len())
+	}
+	if !s1.Has("1") {
+		t.Error(`Expected Has("1") to be true`)
+	}
+	if !s1.Has("2") {
+		t.Error(`Expected Has("2") to be true`)
+	}
+}
+
 func TestClone(t *testing.T) {
 	t.Run("nil", func(t *testing.T) {
 		var s Set[string]