@ash2k 2025-07-26
This panic looks like a legitimate bug? https://prow.k8s.io/view/gs/kubernetes-ci-logs/pr-logs/pull/133231/pull-kubernetes-unit-windows-master/1948977486360481792#1:build-log.txt%3A2478 Fixed in the last commit.