@BenTheElder 2025-07-29
Do we have benchmark results?