@ash2k 2025-07-30
Just pushed a commit with a benchmark. Here is an excerpt with the relevant output:

```
goos: darwin
goarch: arm64
pkg: k8s.io/apimachinery/pkg/util/sets
cpu: Apple M4 Max

BenchmarkStringSet/difference-20
BenchmarkStringSet/difference-20-16                   	  882244	      1335 ns/op	    1636 B/op	       6 allocs/op
BenchmarkStringSet/difference-seq-20
BenchmarkStringSet/difference-seq-20-16               	 2687314	       444.2 ns/op	      56 B/op	       3 allocs/op
BenchmarkStringSet/difference-50
BenchmarkStringSet/difference-50-16                   	  363456	      3267 ns/op	    3457 B/op	       8 allocs/op
BenchmarkStringSet/difference-seq-50
BenchmarkStringSet/difference-seq-50-16               	 1000000	      1164 ns/op	      56 B/op	       3 allocs/op
BenchmarkStringSet/difference-100
BenchmarkStringSet/difference-100-16                  	  190008	      6237 ns/op	    6938 B/op	      10 allocs/op
BenchmarkStringSet/difference-seq-100
BenchmarkStringSet/difference-seq-100-16              	  500174	      2382 ns/op	      56 B/op	       3 allocs/op
BenchmarkStringSet/difference-500
BenchmarkStringSet/difference-500-16                  	   30966	     38250 ns/op	   54276 B/op	      16 allocs/op
BenchmarkStringSet/difference-seq-500
BenchmarkStringSet/difference-seq-500-16              	  118758	      9658 ns/op	      56 B/op	       3 allocs/op
BenchmarkStringSet/difference-1000
BenchmarkStringSet/difference-1000-16                 	   14876	     80534 ns/op	  108803 B/op	      21 allocs/op
BenchmarkStringSet/difference-seq-1000
BenchmarkStringSet/difference-seq-1000-16             	   50556	     23123 ns/op	      56 B/op	       3 allocs/op
```