@BenTheElder 2025-07-30
thanks -- do we have before/after?

this diff also exposes more APIs, some of which seem to be missing tests.