@ash2k 2025-07-30
> thanks -- do we have before/after?

I've added the same test to `master` and ran it:

```
goos: darwin
goarch: arm64
pkg: k8s.io/apimachinery/pkg/util/sets
cpu: Apple M4 Max
BenchmarkStringSet
BenchmarkStringSet/difference-20
BenchmarkStringSet/difference-20-16         	  800041	      1353 ns/op	    1636 B/op	       6 allocs/op
BenchmarkStringSet/difference-50
BenchmarkStringSet/difference-50-16         	  354494	      3314 ns/op	    3457 B/op	       8 allocs/op
BenchmarkStringSet/difference-100
BenchmarkStringSet/difference-100-16        	  183486	      6295 ns/op	    6938 B/op	      10 allocs/op
BenchmarkStringSet/difference-500
BenchmarkStringSet/difference-500-16        	   29953	     39356 ns/op	   54291 B/op	      16 allocs/op
BenchmarkStringSet/difference-1000
BenchmarkStringSet/difference-1000-16       	   14665	     81578 ns/op	  108807 B/op	      21 allocs/op
```

> this diff also exposes more APIs, some of which seem to be missing tests.

- I've added a test for pre-existing `Clone()`
- `DifferenceSeq()` is covered by tests for `Difference()` as the former is used by the later.
- I've added a test for `InsetSet()`.