@ash2k 2025-07-30
To make it easier to see, I've combined the above results below:

```
this PR:

BenchmarkStringSet/difference-20
BenchmarkStringSet/difference-20-16                   	  882244	      1335 ns/op	    1636 B/op	       6 allocs/op
BenchmarkStringSet/difference-50
BenchmarkStringSet/difference-50-16                   	  363456	      3267 ns/op	    3457 B/op	       8 allocs/op
BenchmarkStringSet/difference-100
BenchmarkStringSet/difference-100-16                  	  190008	      6237 ns/op	    6938 B/op	      10 allocs/op
BenchmarkStringSet/difference-500
BenchmarkStringSet/difference-500-16                  	   30966	     38250 ns/op	   54276 B/op	      16 allocs/op
BenchmarkStringSet/difference-1000
BenchmarkStringSet/difference-1000-16                 	   14876	     80534 ns/op	  108803 B/op	      21 allocs/op


master:
BenchmarkStringSet
BenchmarkStringSet/difference-20
BenchmarkStringSet/difference-20-16         	  800041	      1353 ns/op	    1636 B/op	       6 allocs/op
BenchmarkStringSet/difference-50
BenchmarkStringSet/difference-50-16         	  354494	      3314 ns/op	    3457 B/op	       8 allocs/op
BenchmarkStringSet/difference-100
BenchmarkStringSet/difference-100-16        	  183486	      6295 ns/op	    6938 B/op	      10 allocs/op
BenchmarkStringSet/difference-500
BenchmarkStringSet/difference-500-16        	   29953	     39356 ns/op	   54291 B/op	      16 allocs/op
BenchmarkStringSet/difference-1000
BenchmarkStringSet/difference-1000-16       	   14665	     81578 ns/op	  108807 B/op	      21 allocs/op
```